Moving db api things to chainsync/db/api
