--- lib/chainsync/chainsync/db/__init__.py
@@ -1,2 +1,0 @@
-"""Api server for the chainsync database."""
-from .api_server import launch_flask
